requirements.txt
fastapi
uvicorn
pydantic
openai
supabase
python-multipart
python-dotenv
numpy
docling
torch>=2.1.0
transformers>=4.36.0
accelerate